Add robot commands
Each command is a function in the `Robot::Commands` module. All of
these functions have a similar signature: `state`, a `Robot::State`
instance, and `args`, an optional hash. They must return a state.

In case a command needs to update the initial state it will first
duplicate the input state and then return the new state.